chore: forge fmt

+++ evm/src/eNear/contracts/ENearProxy.sol
@@ -2,10 +2,10 @@
 pragma solidity 0.8.24;
 
 import "../../common/Borsh.sol";
-import {AccessControlUpgradeable} from '@openzeppelin/contracts-upgradeable/access/AccessControlUpgradeable.sol';
-import {UUPSUpgradeable} from '@openzeppelin/contracts-upgradeable/proxy/utils/UUPSUpgradeable.sol';
-import {IENear, INearProver} from './IENear.sol';
-import {ICustomMinter} from '../../common/ICustomMinter.sol';
+import {AccessControlUpgradeable} from "@openzeppelin/contracts-upgradeable/access/AccessControlUpgradeable.sol";
+import {UUPSUpgradeable} from "@openzeppelin/contracts-upgradeable/proxy/utils/UUPSUpgradeable.sol";
+import {IENear, INearProver} from "./IENear.sol";
+import {ICustomMinter} from "../../common/ICustomMinter.sol";
 import "../../omni-bridge/contracts/SelectivePausableUpgradable.sol";
 
 contract ENearProxy is UUPSUpgradeable, AccessControlUpgradeable, ICustomMinter, SelectivePausableUpgradable {
@@ -17,14 +17,20 @@ contract ENearProxy is UUPSUpgradeable, AccessControlUpgradeable, ICustomMinter,
     uint256 public currentReceiptId;
     INearProver public prover;
 
-    uint constant PAUSED_LEGACY_FIN_TRANSFER = 1 << 0;
+    uint256 constant PAUSED_LEGACY_FIN_TRANSFER = 1 << 0;
 
     /// @custom:oz-upgrades-unsafe-allow constructor
     constructor() {
         _disableInitializers();
     }
 
-    function initialize(address _eNear, address _prover, bytes memory _nearConnector, uint256 _currentReceiptId, address _adminAddress) public initializer {
+    function initialize(
+        address _eNear,
+        address _prover,
+        bytes memory _nearConnector,
+        uint256 _currentReceiptId,
+        address _adminAddress
+    ) public initializer {
         __UUPSUpgradeable_init();
         __AccessControl_init();
         __Pausable_init();
@@ -58,17 +64,14 @@ contract ENearProxy is UUPSUpgradeable, AccessControlUpgradeable, ICustomMinter,
 
     function burn(address token, uint128 amount) public onlyRole(MINTER_ROLE) {
         require(token == address(eNear), "ERR_INCORRECT_ENEAR_ADDRESS");
-        eNear.transferToNear(amount, string(''));
+        eNear.transferToNear(amount, string(""));
     }
 
-    function finaliseNearToEthTransfer(
-        bytes memory proofData,
-        uint64 proofBlockHeight
-    ) external whenNotPaused(PAUSED_LEGACY_FIN_TRANSFER) {
-        require(
-            prover.proveOutcome(proofData, proofBlockHeight),
-            "Proof should be valid"
-        );
+    function finaliseNearToEthTransfer(bytes memory proofData, uint64 proofBlockHeight)
+        external
+        whenNotPaused(PAUSED_LEGACY_FIN_TRANSFER)
+    {
+        require(prover.proveOutcome(proofData, proofBlockHeight), "Proof should be valid");
 
         eNear.finaliseNearToEthTransfer(proofData, proofBlockHeight);
     }
@@ -77,11 +80,9 @@ contract ENearProxy is UUPSUpgradeable, AccessControlUpgradeable, ICustomMinter,
         _pause(PAUSED_LEGACY_FIN_TRANSFER);
     }
 
-    function pause(uint flags) external onlyRole(DEFAULT_ADMIN_ROLE) {
+    function pause(uint256 flags) external onlyRole(DEFAULT_ADMIN_ROLE) {
         _pause(flags);
     }
 
-    function _authorizeUpgrade(
-        address newImplementation
-    ) internal override onlyRole(DEFAULT_ADMIN_ROLE) {}
+    function _authorizeUpgrade(address newImplementation) internal override onlyRole(DEFAULT_ADMIN_ROLE) {}
 }
