+++ evm/src/eNear/contracts/IENear.sol
@@ -5,8 +5,8 @@ pragma solidity 0.8.24;
 interface IENear {
     function transferToNear(uint256 _amount, string calldata _nearReceiverAccountId) external;
     function finaliseNearToEthTransfer(bytes calldata proofData, uint64 proofBlockHeight) external;
-    function nearConnector() external view returns(bytes memory);
-    function adminSstore(uint key, uint value) external;
+    function nearConnector() external view returns (bytes memory);
+    function adminSstore(uint256 key, uint256 value) external;
     function balanceOf(address account) external view returns (uint256);
     function totalSupply() external view returns (uint256);
     function transfer(address recipient, uint256 amount) external returns (bool);
